requirements.txt
openai
python-dotenv
tenacity
networkx
modelscope
tqdm
